--- Крапивина_Яна.J3112.lab2
@@ -1,40 +1,0 @@
-#include <iostream> 
-#include <unordered_map>
-#include <string>
-#include <vector>
-using namespace std;
-
-// 1 * 10^4 = 40000 байт на вход
-
-int main(){
-    int n;    //4 байта
-    cout << "Введите количество строк в массиве:";    
-    cin >> n;             
-
-
-    vector <string> arr(n);          
-    cout << "Введите строки\n";     
-    for (int i = 0; i < n; ++i){     //4*50 = 200 байт, 50 - средняя длина строки
-        cin >> arr[i];       
-    }
-
-
-    unordered_map<string, int> countMap;     //4 * 50 = 200 байт
-
-    for (const auto& str : arr){
-        countMap[str]++;
-    }
-
-    for (string& str : arr){
-        str = to_string(countMap[str]);
-    }
-
-    for (const auto& str : arr){
-        cout << str << endl;
-    }
-
-    return 0; 
-}
-
-
-// 4 + 200 + 200 = 404 байта
